+++ packages/app/src/utils/agent-snapshots.ts
@@ -2,6 +2,7 @@ import type { AgentSnapshotPayload } from "@getpaseo/protocol/messages";
 import type { AgentPermissionRequest } from "@getpaseo/protocol/agent-types";
 import { getParentAgentIdFromLabels } from "@getpaseo/protocol/agent-labels";
 import type { ActiveTurnIdentity } from "@/timeline/turn-liveness";
+import type { Agent } from "@/stores/session-store";
 
 function normalizeActiveTurn(
   snapshot: AgentSnapshotPayload,
@@ -17,6 +18,17 @@ function normalizeActiveTurn(
   return snapshot.status === "running" ? { turnId: null, startedAt: lastUserMessageAt } : null;
 }
 
+function projectActiveTurn(agent: Agent): Pick<AgentSnapshotPayload, "activeTurn"> {
+  if (agent.activeTurn === null) return { activeTurn: null };
+  if (agent.activeTurn.turnId === null) return {};
+  return {
+    activeTurn: {
+      turnId: agent.activeTurn.turnId,
+      startedAt: agent.activeTurn.startedAt?.toISOString() ?? null,
+    },
+  };
+}
+
 export function derivePendingPermissionKey(
   agentId: string,
   request: AgentPermissionRequest,
@@ -29,6 +41,37 @@ export function derivePendingPermissionKey(
     `${request.kind}:${JSON.stringify(request.input ?? request.metadata ?? {})}`;
 
   return `${agentId}:${fallbackId}`;
+}
+
+export function projectAgentSnapshot(agent: Agent): AgentSnapshotPayload {
+  return {
+    id: agent.id,
+    provider: agent.provider,
+    cwd: agent.cwd,
+    ...(agent.workspaceId ? { workspaceId: agent.workspaceId } : {}),
+    model: agent.model,
+    ...(agent.features ? { features: agent.features } : {}),
+    thinkingOptionId: agent.thinkingOptionId ?? null,
+    createdAt: agent.createdAt.toISOString(),
+    updatedAt: agent.updatedAt.toISOString(),
+    lastUserMessageAt: agent.lastUserMessageAt?.toISOString() ?? null,
+    status: agent.status,
+    ...projectActiveTurn(agent),
+    capabilities: agent.capabilities,
+    currentModeId: agent.currentModeId,
+    availableModes: agent.availableModes,
+    pendingPermissions: agent.pendingPermissions,
+    persistence: agent.persistence,
+    ...(agent.runtimeInfo ? { runtimeInfo: agent.runtimeInfo } : {}),
+    ...(agent.lastUsage ? { lastUsage: agent.lastUsage } : {}),
+    ...(agent.lastError ? { lastError: agent.lastError } : {}),
+    title: agent.title,
+    labels: agent.labels,
+    requiresAttention: agent.requiresAttention ?? false,
+    attentionReason: agent.attentionReason ?? null,
+    attentionTimestamp: agent.attentionTimestamp?.toISOString() ?? null,
+    archivedAt: agent.archivedAt?.toISOString() ?? null,
+  };
 }
 
 export function normalizeAgentSnapshot(snapshot: AgentSnapshotPayload, serverId: string) {
